rag_app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
openai==1.10.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
pypdf==4.0.1
rank-bm25==0.2.2
python-dotenv==1.0.0
httpx==0.26.0
streamlit
